erp-backend/requirements.txt
# Core
Django==5.2.5
djangorestframework==3.14.0
dj-database-url==2.1.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# CORS & Security
django-cors-headers==4.3.1
whitenoise==6.6.0

# API Documentation
drf-yasg==1.21.7
django-filter==23.3

# Data processing
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2

# Redis
django-redis==6.2.0

# Gunicorn
gunicorn==21.2.0
